refactor(calendar) Rename CalendarHeading to MonthViewSelect
